+++ backend/src/main/java/reviewme/security/aspect/ReviewAuthorizationAspect.java
@@ -1,7 +1,7 @@
-package reviewme.global.authorization;
+package reviewme.security.aspect;
 
-import static reviewme.global.authorization.ResourceAuthorizationUtils.getCurrentSession;
-import static reviewme.global.authorization.ResourceAuthorizationUtils.getTarget;
+import static reviewme.security.aspect.ResourceAuthorizationUtils.getCurrentSession;
+import static reviewme.security.aspect.ResourceAuthorizationUtils.getTarget;
 
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -10,14 +10,14 @@ import org.aspectj.lang.annotation.Around;
 import org.aspectj.lang.annotation.Aspect;
 import org.springframework.stereotype.Component;
 import reviewme.auth.domain.GitHubMember;
-import reviewme.global.authorization.exception.ReviewGroupNotExistsBySessionReviewRequestCodeException;
-import reviewme.global.authorization.exception.ForbiddenReviewAccessException;
-import reviewme.global.session.SessionManager;
+import reviewme.security.aspect.exception.ForbiddenReviewAccessException;
+import reviewme.security.aspect.exception.ReviewGroupNotExistsBySessionReviewRequestCodeException;
 import reviewme.review.domain.Review;
 import reviewme.review.repository.ReviewRepository;
 import reviewme.review.service.exception.ReviewNotFoundException;
 import reviewme.reviewgroup.domain.ReviewGroup;
 import reviewme.reviewgroup.repository.ReviewGroupRepository;
+import reviewme.security.session.SessionManager;
 
 @Aspect
 @Component
