+++ backend/src/test/java/reviewme/global/authorization/AopTestClass.java
@@ -1,6 +1,8 @@
 package reviewme.global.authorization;
 
 import org.springframework.stereotype.Component;
+import reviewme.security.aspect.RequireReviewAccess;
+import reviewme.security.aspect.RequireReviewGroupAccess;
 
 @Component
 class AopTestClass {
