+++ backend/src/test/java/reviewme/global/authorization/ResourceAuthorizationUtilsTest.java
@@ -15,7 +15,8 @@ import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpSession;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
-import reviewme.global.authorization.exception.SpELEvaluationFailedException;
+import reviewme.security.aspect.ResourceAuthorizationUtils;
+import reviewme.security.aspect.exception.SpELEvaluationFailedException;
 import reviewme.support.ServiceTest;
 
 @ServiceTest
